requirements.txt
# Core dependencies per PaddleOCR-VL API Server
# Compatible con CUDA 12.1 e Google Cloud Run GPU L4

# PyTorch (installato via index-url nel Dockerfile)
# torch==2.1.0
# torchvision==0.16.0

# Transformers e accelerazione
transformers>=4.45.0
accelerate>=0.25.0
einops>=0.7.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# Web server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Image processing
pillow>=10.0.0
opencv-python-headless>=4.8.0

# PDF processing (license-compliant: no GPL/AGPL)
pypdfium2>=4.26.0
pikepdf>=8.10.0
pdf2image>=1.16.3

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.2.0
